+++ app/src/main/java/com/example/hikeculator/domain/entities/NutritionalValue.kt
@@ -1,8 +1,8 @@
 package com.example.hikeculator.domain.entities
 
 data class NutritionalValue(
-    val calories: Long,
-    val proteins: Long,
-    val fats: Long,
-    val carbohydrates: Long
+    val calories: Double,
+    val proteins: Double,
+    val fats: Double,
+    val carbohydrates: Double,
 )
