+++ app/src/main/java/com/example/hikeculator/domain/entities/Product.kt
@@ -2,6 +2,6 @@ package com.example.hikeculator.domain.entities
 
 data class Product(
     val name: String,
-    val weight: Long,
+    val weight: Double,
     val nutritionalValue: NutritionalValue,
 )
